+++ app/snappymail/v/2.38.2/app/libraries/snappymail/gpg/smime.php
@@ -43,7 +43,7 @@ class SMIME extends Base
 		// How to use gpgme-json ?
 		$this->binary = static::findBinary('gpgsm');
 
-		$info = \preg_replace('/\R +/', ' ', `$this->binary --version`);
+		$info = $this->binary ? \preg_replace('/\R +/', ' ', (string) \shell_exec(\escapeshellarg($this->binary) . ' --version')) : '';
 		if (\preg_match('/gpgsm.+([0-9]+\\.[0-9]+\\.[0-9]+)/', $info, $match)) {
 			$this->version = $match[1];
 		}
@@ -589,7 +589,7 @@ class SMIME extends Base
 	/**
 	 * Verifies a signed text
 	 */
-	public function verify(string $signed_text, string $signature, string &$plaintext = null) /*: array|false*/
+	public function verify(string $signed_text, string $signature, ?string &$plaintext = null) /*: array|false*/
 	{
 		return $this->_verify($signed_text, $signature);
 	}
@@ -597,7 +597,7 @@ class SMIME extends Base
 	/**
 	 * Verifies a signed file
 	 */
-	public function verifyFile(string $filename, string $signature, string &$plaintext = null) /*: array|false*/
+	public function verifyFile(string $filename, string $signature, ?string &$plaintext = null) /*: array|false*/
 	{
 		$fp = \fopen($filename, 'rb');
 		try {
@@ -613,7 +613,7 @@ class SMIME extends Base
 	/**
 	 * Verifies a signed file
 	 */
-	public function verifyStream($fp, string $signature, string &$plaintext = null) /*: array|false*/
+	public function verifyStream($fp, string $signature, ?string &$plaintext = null) /*: array|false*/
 	{
 		if (!$fp || !\is_resource($fp)) {
 			throw new \Exception('Invalid stream resource');